Refactor imports to use data.js instead of products.js in order summary, payment summary, and view details scripts
